🔨 refactor: FormCard 컴포넌트 명 NeogaFormTicket으로 변경
