+++ src/presentation/components/common/NeogaFormTicket/style.ts
